requirements.txt
aiohttp[speedups]~=3.6.2
flask~=1.1.0
gunicorn~=20.0.0
starlette~=0.13.4
uvicorn~=0.11.5
uwsgi~=2.0.0
aiopg~=1.0.0
psycopg2-binary~=2.8.5
gevent~=20.5.0
meinheld~=1.0.1
falcon~=2.0.0
sanic~=19.12.2
bottle~=0.12.18
daphne~=2.5.0
Quart==0.13.0
requests-2.24.0
